(feat) Make all nodes show up in a fairly nice place
This was done by implementing a fixXCoords function, which will check
if the x-coordinate of each node is extremely high and change it if
it is. Could be implemented much better.

Potentially the entire dndTree.json file should probably be refactored
heavily to intorduce the functionality we want; otherwise it feels like
we will be stuck fixing patches once we start dynamically rendering
the trees.